refactor: move InvokableOperation
