+++ basyx.submodelservice/basyx.submodelservice-core/src/main/java/org/eclipse/digitaltwin/basyx/operation/InvokableOperation.java
@@ -24,7 +24,7 @@
  ******************************************************************************/
 
 
-package org.eclipse.digitaltwin.basyx;
+package org.eclipse.digitaltwin.basyx.operation;
 
 import java.util.function.Function;
